- Added a new function 'btype' to handle %b conversion - Implemented 'decimalToBinary' function to convert decimal to binary - Updated _printf to recognize and process the %b specifier - Modified main.c to include a test case for %b
