+++ main.h
@@ -13,14 +13,14 @@ int _putchar(char c);
 int ft_strlen(char *str);
 int convert_to_string(int num, char *str);
 int find_format_type(char format_char, print_specifiers *specifiers);
-
-
+int decimalToBinary(int n, char *str);
 
 int ctype(va_list args);
 int stype(va_list args);
 int mtype(va_list args);
 int dtype(va_list args);
 int itype(va_list args);
+int btype(va_list args);
 
 int _printf(const char *format, ...);
